Code/requirement.txt
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
filelock==3.13.1
fonttools==4.53.1
fsspec==2024.2.0
idna==3.6
imageio==2.33.1
Jinja2==3.1.3
kiwisolver==1.4.5
lazy_loader==0.3
MarkupSafe==2.1.5
matplotlib==3.9.1
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
packaging==23.2
pillow==10.2.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
requests==2.31.0
scikit-image==0.22.0
scikit-learn==1.5.1
scipy==1.12.0
six==1.16.0
sympy==1.12
thop==0.1.1.post2209072238
tifffile==2024.1.30
torch==2.2.0
torchvision==0.17.0
triton==2.2.0
typing_extensions==4.9.0
urllib3==2.2.0
notebook